projects/publication-assistant/requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyGithub==2.1.1
yake==0.4.8
textstat==0.7.3
pydantic==2.5.0
typing-extensions==4.8.0
